+++ config/config.go
@@ -3,13 +3,12 @@ package config
 import "time"
 
 type Config struct {
-	Proxy   Proxy
-	Listen  []Listen
-	Routes  string
-	Metrics []Metrics
-	Consul  Consul
-	UI      UI
-	Runtime Runtime
+	Proxy    Proxy
+	Registry Registry
+	Listen   []Listen
+	Metrics  []Metrics
+	UI       UI
+	Runtime  Runtime
 }
 
 type Listen struct {
@@ -51,6 +50,21 @@ type Metrics struct {
 	Prefix   string
 	Interval time.Duration
 	Addr     string
+}
+
+type Registry struct {
+	Backend string
+	Static  Static
+	File    File
+	Consul  Consul
+}
+
+type Static struct {
+	Routes string
+}
+
+type File struct {
+	Path string
 }
 
 type Consul struct {